pyproject.toml
[tool.poetry]
name = "nonebot_plugin_simplemusic"
version = "0.1.0"
description = "适用于 Nonebot2 的点歌插件"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/MeetWq/nonebot-plugin-simplemusic"
repository = "https://github.com/MeetWq/nonebot-plugin-simplemusic"

[tool.poetry.dependencies]
python = "^3.7.3"
httpx = ">=0.19.0"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
